refactor: improve code organization and documentation
