+++ src/features/interactive-cube/hooks/useIsSpacePressed.ts
@@ -1,6 +1,15 @@
+/**
+ * Hook for tracking space key press state.
+ * 
+ * @module hooks/useIsSpacePressed
+ */
+
 import { useEffect, useState } from "react";
 
-export function useIsSpacePressed() {
+/**
+ * Tracks whether the space key is currently pressed.
+ */
+export function useIsSpacePressed(): boolean {
   const [isSpacePressed, setIsSpacePressed] = useState(false);
 
   useEffect(() => {
@@ -10,18 +19,18 @@ export function useIsSpacePressed() {
       }
     };
 
-    const handleKyeUp = (e: KeyboardEvent) => {
+    const handleKeyUp = (e: KeyboardEvent) => {
       if (e.code === "Space" && isSpacePressed) {
         setIsSpacePressed(false);
       }
     };
 
     document.addEventListener("keydown", handleKeyDown);
-    document.addEventListener("keyup", handleKyeUp);
+    document.addEventListener("keyup", handleKeyUp);
 
     return () => {
       document.removeEventListener("keydown", handleKeyDown);
-      document.removeEventListener("keyup", handleKyeUp);
+      document.removeEventListener("keyup", handleKeyUp);
     };
   }, [isSpacePressed]);
