+++ src/features/interactive-cube/logic/cube-state.ts
@@ -1,11 +1,18 @@
+/**
+ * Cube state model and factory functions.
+ * 
+ * @module logic/cube-state
+ */
+
 export type Face = "U" | "D" | "L" | "R" | "F" | "B";
-
 export type Sticker = "white" | "yellow" | "red" | "orange" | "green" | "blue";
-
 export type CubeState = Record<Face, Sticker[]>;
 
+/**
+ * Creates a solved cube state with standard color scheme.
+ */
 export function createSolvedState(): CubeState {
-  const state = {
+  return {
     U: Array(9).fill("yellow"),
     D: Array(9).fill("white"),
     L: Array(9).fill("red"),
@@ -13,6 +20,4 @@ export function createSolvedState(): CubeState {
     F: Array(9).fill("green"),
     B: Array(9).fill("blue"),
   };
-
-  return state;
 }
