--- src/features/interactive-cube/types.ts
@@ -1,14 +1,0 @@
-import * as THREE from "three";
-
-export interface PegatineTextures {
-  red: THREE.Texture;
-  orange: THREE.Texture;
-  blue: THREE.Texture;
-  green: THREE.Texture;
-  white: THREE.Texture;
-  yellow: THREE.Texture;
-}
-
-export type PegatineColor = keyof PegatineTextures;
-
-
